requirements.txt
uvicorn==0.11.5
fastapi==0.58.0
aioredis==1.3.1
async-caches==0.3.0
fastapi-plugins==0.4.1
